requirements.txt
Pillow==10.4.0
torch==2.4.0
matplotlib
facexlib
streamlit
huggingface_hub
opencv-python-headless
